+++ modal-go/test/sandbox_test.go
@@ -37,7 +37,7 @@ func TestCreateOneSandboxTopLevelImageAPI(t *testing.T) {
 	app, err := modal.AppLookup(context.Background(), "libmodal-test", &modal.LookupOptions{CreateIfMissing: true})
 	g.Expect(err).ShouldNot(gomega.HaveOccurred())
 
-	image := modal.NewImageFromRawRegistry("alpine:3.21", nil)
+	image := modal.NewImageFromRegistry("alpine:3.21", nil)
 	g.Expect(image.ImageId).Should(gomega.BeEmpty())
 
 	sb, err := app.CreateSandbox(image, nil)
@@ -58,7 +58,7 @@ func TestCreateOneSandboxTopLevelImageAPISecret(t *testing.T) {
 	})
 	g.Expect(err).ShouldNot(gomega.HaveOccurred())
 
-	image := modal.NewImageFromRawRegistry("us-east1-docker.pkg.dev/modal-prod-367916/private-repo-test/my-image", &modal.ImageFromRegistryOptions{
+	image := modal.NewImageFromRegistry("us-east1-docker.pkg.dev/modal-prod-367916/private-repo-test/my-image", &modal.ImageFromRegistryOptions{
 		Secret: secret,
 	})
 	g.Expect(image.ImageId).Should(gomega.BeEmpty())