+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Both client libraries are pre-1.0, and they have separate versioning.
   - `proxy`: Connect a Modal Proxy to a Sandbox.
   - `workdir`: Set the working directory.
 - Add top level for Image objects that are lazy. The images are built when creating a sandbox.
-  - `Image.FromRawRegistry` in typescript and `NewImageFromRawRegistry` in golang.
+  - `Image.FromRegistry` in typescript and `NewImageFromRegistry` in golang.
   - `Image.ImageFromAwsEcrTopLevel` in typescript and `NewImageImageFromAwsEcrTopLevel` in golang.
   - `Image.FromGcpArtifactRegistry` in typescript and `NewImageFromGcpArtifactRegistry` in golang.